[libnozzle] add nozzle_get_mac / nozzle_set_mac / nozzle_reset_mac tests
